Refactor occupancy map generation
